requirements.txt
Django==4.2
boto3==1.34.0
python-dotenv==1.0.0
